pyproject.toml
[project]
name = "fastopp-site"
version = "1.0.0"
description = "Marketing site for FastOpp - FastAPI starter package for AI web applications"
authors = [
    {name = "Oppkey", email = "info@oppkey.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "sqladmin>=0.15.0",
    "alembic>=1.12.0",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "aiofiles>=23.0.0",
    "pillow>=10.0.0",
    "aiosqlite>=0.19.0",
    "sqlmodel>=0.0.8",
    "fastapi-users[sqlalchemy]>=12.0.0",
    "requests>=2.31.0",
    "sse-starlette>=1.8.0",
    "aiohttp>=3.9.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
